requirements.txt
python-dotenv
aiogram
httpx
